requirements.txt
pybind11>=2.6.0
pandas
matplotlib
seaborn
Pillow